:smile:1、整理前端代码，配置ajax全局setting 2、捕获数据库相关异常 3、新增商品分类

+++ src/main/java/com/example/demo/mapper/CategoryMapper.java
@@ -10,7 +10,7 @@ public interface CategoryMapper {
 
     int deleteByExample(CategoryExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Category record);
 
@@ -18,7 +18,7 @@ public interface CategoryMapper {
 
     List<Category> selectByExample(CategoryExample example);
 
-    Category selectByPrimaryKey(Integer id);
+    Category selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Category record, @Param("example") CategoryExample example);
